backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
flask-compress>=1.15.0
flask-sqlalchemy>=3.1.1
pyjwt>=2.8.0
gunicorn>=21.2.0
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0
shapely>=2.0.0
orjson>=3.10.0